Move StaticPage component
