+++ src/containers/PageBuilder/StaticPage.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import { node } from 'prop-types';
 import { connect } from 'react-redux';
-import { isScrollingDisabled } from '../../ducks/UI.duck';
-import { Page } from '../../components';
+
+import { isScrollingDisabled } from '../../ducks/UI.duck.js';
+import { Page } from '../../components/index.js';
 
 const StaticPageComponent = props => {
   const { children, ...pageProps } = props;
