--- src/containers/StaticPage/README.md
@@ -1,4 +1,0 @@
-# StaticPage
-
-Component for creating static pages. See the
-[Static pages documentation](../../../docs/static-pages.md) for more information.
