+++ src/containers/index.js
@@ -1,3 +1,3 @@
 export { default as NotFoundPage } from './NotFoundPage/NotFoundPage';
-export { default as StaticPage } from './StaticPage/StaticPage';
+export { default as StaticPage } from './PageBuilder/StaticPage';
 export { default as TopbarContainer } from './TopbarContainer/TopbarContainer';
